+++ src/lib.rs
@@ -195,6 +195,7 @@ pub mod objc;
 
 pub mod app_kit;
 pub mod core_animation;
+pub mod core_audio;
 pub mod core_foundation;
 pub mod core_graphics;
 pub mod core_image;